Unit 1, Lesson 3.3 Cleaning Attempt 2
Implemented some of the changes we discussed in today's session (10/10. There are a few bugs that I'd appreciate help with. I will also continue to look into the bag-of-words model to extract topics from the set of journal titles. I'd like to know whether the approach I described in the final markdown cell would work.