Unit 2, Lesson 2.7 First draft
It's an okay start, following the model in the lesson, but I'm having trouble with the code to implement the NLP tools I'd like to try out to improve the model. My questions about this will mostly have to do with getting the code right and understanding the error messages. I've spent a lot of time on StackExchange and in the Python documentation, but I'm getting stuck.